rename views and adapt routes
